Basilisk II compiles on Xcode 8, as a 64-bit Mac app
SDL 1.x is used for display, rather than Mac OS X specific backend.  If time permits, I'll port it to SDL 2, if only to reduce Basilisk's overall code foot-print.

Lots of features are apt to be disabled, as many 'dummy' backends were used.

Video-depths other than 1-bit or 32-bit are untested, and in some cases (4-bit, at least) are currently non-functional.  This is due to a partial re-write of the SDL backend's blitting code, which was non-functional when low-bit-depths were used.

The SDL backend was also rewired, on OSX, to not attempt to align the display buffer on page-boundaries.  So far, this doesn't seem to cause any notice-able problems, however, that's only using limited knowledge and testing (System 7.5.x does boot and display at 640x480, though!).  The original display-buffer allocation code was failing to run, in some cases.

Preferences are, on Mac, currently hardcoded to be accessed at /tmp/BasiliskII/BasiliskII_Prefs.  The folder, "/tmp/BasiliskII/", may be a symbolic link to elsewhere, though.

+++ BasiliskII/src/uae_cpu/gencpu.c
@@ -2189,14 +2189,14 @@ static void gen_opcode (unsigned long int opcode)
      case i_CINVA:
 	/* gb-- srcreg now contains the cache field */
 	printf ("\tif (srcreg&0x2)\n");
-	printf ("\t\tflush_icache(%d);\n", 30 + ((opcode >> 3) & 3)); 
+	printf ("\t\tflush_icache(%d);\n", (int)(30 + ((opcode >> 3) & 3))); 
 	break;
      case i_CPUSHL:
      case i_CPUSHP:
      case i_CPUSHA:
 	/* gb-- srcreg now contains the cache field */
 	printf ("\tif (srcreg&0x2)\n");
-	printf ("\t\tflush_icache(%d);\n", 40 + ((opcode >> 3) & 3)); 
+	printf ("\t\tflush_icache(%d);\n", (int)(40 + ((opcode >> 3) & 3))); 
 	break;
      case i_MOVE16:
 	if ((opcode & 0xfff8) == 0xf620) {
